requirements.txt
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
docopt==0.6.2
Flask==3.0.0
google-api-core==2.12.0
google-api-python-client==2.103.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
keyboard==0.13.5
MarkupSafe==2.1.3
numpy==1.26.1
oauthlib==3.2.2
opencv-python==4.8.1.78
Pillow==10.1.0
pipreqs==0.4.13
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pynput==1.7.6
pyparsing==3.1.1
PySimpleGUI==4.60.5
pywin32==306
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==2.0.6
Werkzeug==3.0.0
WMI==1.5.1
yarg==0.1.9
